Feature/add module file path finder (#2)
* Update mail_merge_scheduler.py

Update to find the file path of the module and prepend it to all files being referenced in the script. This allows more flexibility and prepares the library for packaging as a site-package for pip installation.

* Update schedules.py

Update to find the file path of the module and prepend it to all files being referenced in the script. This allows more flexibility and prepares the library for packaging as a site-package for pip installation.