Update mail_merge_scheduler.py
The generate_list_of_next_days method was an old method, that did not work, however instead of replacing it with the more recent, working version it, the method has been completely reworked to be more efficient, easier to understand, and not unnecessarily complex. 

The reworked version is an all around improvement.

Also some various corrections to overrun lines.