Feature/ major reworking of scheduling to reduce redundant task calls (#8)
* Update mail_merge_scheduler.py

This update eliminates redundant calls from Windows Task Scheduler to schedules.py, by allowing scheduled merges to share a task name. This update should now make this code library far more scalable.

- Added a new attribute, task_name, which represents the name of the task in Windows Task Scheduler, for that particular scheduled mail merge. This attribute is also written to the scheduled_merges.ini file.
- Added 2 new methods, find_tasks_with_same_schedule and generate_task_name.
- Moved error checking from the __init__ method into it's own method, to improve readability.
- Two new imports, ast, and dateutil.parser.
- Moved some things around to be more coherent.
- Minor changes to some doc-strings/comments.

* Update schedules.py

Minor spacing changes.

* Rename weekly_xml_schedule_template.xml to xml_schedule_template.xml

Changed the name of the xml file, because it was unnecessarily long.

* Update mail_merge_scheduler.py

Added the short-hand for the names of the days in self.sched_days to the task_name string. This should make searching for the right task in Windows Task Scheduler much easier.

* Update readme.rst

Updated the Readme to reflect the changes added in resent feature-branches.