Development branch (#9)
* Feature/add module file path finder (#2)

* Update mail_merge_scheduler.py

Update to find the file path of the module and prepend it to all files being referenced in the script. This allows more flexibility and prepares the library for packaging as a site-package for pip installation.

* Update schedules.py

Update to find the file path of the module and prepend it to all files being referenced in the script. This allows more flexibility and prepares the library for packaging as a site-package for pip installation.

* Feature/rework xml gen method in mail merge scheduler.py (#3)

* Update schedules.py

Some bug-fixes with white space, and various changes to improve readability, and get rid of unused variables.

* Update mail_merge_scheduler.py

Reworked the xml_gen method, using BeautifulSoup4 instead of lxml, however the final xml must be overwritten with lxml to change the encoding to be compatible with Windows Task Scheduler.
- Moved the code for finding the windowless path to its own method to make the xml_gen method less cluttered, and to improve readability.
- Various changes to get rid of unused variable names, and make things more readable.

* Update weekly_xml_schedule_template.xml

Reworked the template xml file to work with the new xml_gen method

* Update mail_merge_scheduler.py

Added comment for clarification in the new xml_gen method.

* Feature/major re work of the config file (#4)

* Update mail_merge_scheduler.py

- Updated multiple methods to follow the new structure when writing data to the config file.
- Corrected some misspellings, grammar, and white-space issues in doc-strings and comments.

* Update schedules.py

Major reworking of this file, changes include:
- Removed the populate_from_dict method, and replaced it by instantiating the object with all the variables.
- Reworked multiple methods to properly read from, and write to the config file, following the new structure for the config file.
- Renamed some method and variable names to be more consistent.
- Corrected some misspellings, grammar, and white-space issues in doc-strings and comments.

* Update scheduled_merges.ini

Update config file to reflect the new config structure.

* Update mail_merge_scheduler.py (#5)

The generate_list_of_next_days method was an old method, that did not work, however instead of replacing it with the more recent, working version it, the method has been completely reworked to be more efficient, easier to understand, and not unnecessarily complex. 

The reworked version is an all around improvement.

Also some various corrections to overrun lines.

* Feature/ major reworking of scheduling to reduce redundant task calls (#8)

* Update mail_merge_scheduler.py

This update eliminates redundant calls from Windows Task Scheduler to schedules.py, by allowing scheduled merges to share a task name. This update should now make this code library far more scalable.

- Added a new attribute, task_name, which represents the name of the task in Windows Task Scheduler, for that particular scheduled mail merge. This attribute is also written to the scheduled_merges.ini file.
- Added 2 new methods, find_tasks_with_same_schedule and generate_task_name.
- Moved error checking from the __init__ method into it's own method, to improve readability.
- Two new imports, ast, and dateutil.parser.
- Moved some things around to be more coherent.
- Minor changes to some doc-strings/comments.

* Update schedules.py

Minor spacing changes.

* Rename weekly_xml_schedule_template.xml to xml_schedule_template.xml

Changed the name of the xml file, because it was unnecessarily long.

* Update mail_merge_scheduler.py

Added the short-hand for the names of the days in self.sched_days to the task_name string. This should make searching for the right task in Windows Task Scheduler much easier.

* Update readme.rst

Updated the Readme to reflect the changes added in resent feature-branches.